requirements.txt
flask==2.0.1
openai-whisper==20231117
gunicorn==20.1.0
werkzeug==2.0.0
flask-cors==3.0.10
